Move obfuscation out of Redactor abstraction
