+++ trace/obfuscation_test.go
@@ -1,0 +1,11 @@
+package trace
+
+import "testing"
+
+func BenchmarkObfuscation(b *testing.B) {
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		ObfuscateMethod(testWitness.Method)
+	}
+}
