+++ trace/redactor.go
@@ -43,41 +43,6 @@ func NewRedactor() *Redactor {
 	}
 }
 
-// Replaces all primitive values in the given method with zero values.
-func (o *Redactor) ZeroAllPrimitivesInMethod(m *pb.Method) {
-	var ov zeroPrimitivesVisitor
-	vis.Apply(&ov, m)
-
-	// Mark the method as obfuscated.
-	m.GetMeta().GetHttp().Obfuscation = pb.HTTPMethodMeta_ZERO_VALUE
-}
-
-// Replaces all primitive values with zero values.
-type zeroPrimitivesVisitor struct {
-	vis.DefaultSpecVisitorImpl
-}
-
-var _ vis.DefaultSpecVisitor = (*zeroPrimitivesVisitor)(nil)
-
-// EnterData processes the given data and replaces all the primitive values
-// with zero values, regardless of its metadata.
-func (*zeroPrimitivesVisitor) EnterData(self interface{}, _ vis.SpecVisitorContext, d *pb.Data) Cont {
-	dp := d.GetPrimitive()
-	if dp == nil {
-		return Continue
-	}
-
-	pv, err := spec_util.PrimitiveValueFromProto(dp)
-	if err != nil {
-		printer.Warningf("failed to zero out raw value, dropping\n")
-		d.Value = nil
-		return Continue
-	}
-
-	dp.Value = pv.Obfuscate().ToProto().Value
-	return Continue
-}
-
 func (o *Redactor) RedactSensitiveData(m *pb.Method) {
 	pov := redactSensitiveInfoVisitor{
 		redactionOptions: o,
